Abstracted db connection pool out of route handlers. Moved route handler logic for todos into controllers module.
